add test for all env configs; make all envs validate
